corrected the WaitList tables and queries
replaced the bigint with int , added procedures for geting the users from a waitlist or for geting all the waitlists assigned to a user. Change the connection with Contract for endDate to Product endDate

+++ ArtAttack/Database/Entities/UserWaitList.sql
@@ -1,8 +1,8 @@
 create table [UserWaitList]
 (
-	[UserWaitListID] bigint identity primary key,
+	[UserWaitListID] int identity primary key,
 
-	[productWaitListID] bigint not null foreign key references [WaitListProduct]([WaitListProductID]),
+	[productWaitListID] int not null foreign key references [WaitListProduct]([WaitListProductID]),
 
 	[userID] integer not null foreign key references [DummyBuyer]([ID]),
